+++ src/configure_settings/configure_settings.jl
@@ -24,6 +24,7 @@ function default_settings()
         "MultiStage" => 0,
         "MethodofMorris" => 0,
         "IncludeLossesInESR" => 0,
+        "HydrogenMimimumProduction" => 0,
         "EnableJuMPStringNames" => false,
         "HydrogenHourlyMatching" => 0,
         "DC_OPF" => 0,