+++ src/multi_stage/configure_multi_stage_inputs.jl
@@ -136,6 +136,11 @@ function configure_multi_stage_inputs(inputs_d::Dict,
                 inv_cost_wind_per_mwyr.(gen_VRE_STOR),
                 capital_recovery_period_wind.(gen_VRE_STOR),
                 tech_wacc_wind.(gen_VRE_STOR))
+            gen_VRE_STOR.inv_cost_elec_per_mwyr = compute_overnight_capital_cost(
+                settings_d,
+                inv_cost_elec_per_mwyr.(gen_VRE_STOR),
+                capital_recovery_period_elec.(gen_VRE_STOR),
+                tech_wacc_elec.(gen_VRE_STOR))
             gen_VRE_STOR.inv_cost_discharge_dc_per_mwyr = compute_overnight_capital_cost(
                 settings_d,
                 inv_cost_discharge_dc_per_mwyr.(gen_VRE_STOR),
@@ -163,6 +168,8 @@ function configure_multi_stage_inputs(inputs_d::Dict,
                                                         OPEXMULT
             gen_VRE_STOR.fixed_om_wind_cost_per_mwyr = fixed_om_wind_cost_per_mwyr.(gen_VRE_STOR) .*
                                                        OPEXMULT
+            gen_VRE_STOR.fixed_om_elec_cost_per_mwyr = fixed_om_elec_cost_per_mwyr.(gen_VRE_STOR) .*
+                                                       OPEXMULT
             gen_VRE_STOR.fixed_om_cost_discharge_dc_per_mwyr = fixed_om_cost_discharge_dc_per_mwyr.(gen_VRE_STOR) .*
                                                                OPEXMULT
             gen_VRE_STOR.fixed_om_cost_charge_dc_per_mwyr = fixed_om_cost_charge_dc_per_mwyr.(gen_VRE_STOR) .*
@@ -176,7 +183,6 @@ function configure_multi_stage_inputs(inputs_d::Dict,
 
     retirable = is_retirable(gen)
 
-    # TODO: ask Sam about this
     # Set of all resources eligible for capacity retirements
     inputs_d["RET_CAP"] = retirable
     # Set of all storage resources eligible for energy capacity retirements
@@ -188,6 +194,7 @@ function configure_multi_stage_inputs(inputs_d::Dict,
         inputs_d["RET_CAP_DC"] = intersect(retirable, inputs_d["VS_DC"])
         inputs_d["RET_CAP_SOLAR"] = intersect(retirable, inputs_d["VS_SOLAR"])
         inputs_d["RET_CAP_WIND"] = intersect(retirable, inputs_d["VS_WIND"])
+        inputs_d["RET_CAP_ELEC"] = intersect(retirable, inputs_d["VS_ELEC"])
         inputs_d["RET_CAP_STOR"] = intersect(retirable, inputs_d["VS_STOR"])
         inputs_d["RET_CAP_DISCHARGE_DC"] = intersect(retirable,
             inputs_d["VS_ASYM_DC_DISCHARGE"])