+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added default settings in multitage optimization (#703)
 - Added the option to write results after each iteration of a multistage run (Myopic) (#704)
 - Add option to precompile `run_genx_case!` and `cluster_inputs` to speedup TTFX of GenX (#718)
+- Added co-location of electrolyzers to VRE_STOR and eneabled regional hydrogen production limits (#678)
 
 ## [0.4.0] - 2024-03-18
 
@@ -55,7 +56,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   non-zero in multi-stage GenX (#666)
 - Added condition number scaling added to objective function (#667)
 - Added versioned doc-pages for v0.3.6 and v0.4.0
-
 - Added a warning message in write_costs_multistage mentioning th approximate value of costs currently.
 
 ### Fixed